[step3] refactor : 추가된 Score 팩토리 메서드를 테스트 코드에 적용
